Added completed CRUD.

+++ Models/Product.swift
@@ -12,12 +12,12 @@ import SwiftData
 final class Product {
     
     var name: String
-    var tax: Int
+    var tax: Double
     var price: Double
     var isPaidWithCreditCard: Bool
     var photo: String
     
-    init(name: String, tax: Int, price: Double, isPaidWithCreditCard: Bool, photo: String) {
+    init(name: String, tax: Double, price: Double, isPaidWithCreditCard: Bool, photo: String) {
         self.name = name
         self.tax = tax
         self.price = price
